+++ sentry-spring/src/main/java/io/sentry/spring/tracing/SentrySpan.java
@@ -27,7 +27,7 @@ public @interface SentrySpan {
    *
    * @return operation.
    */
-  String op() default "";
+  String operation() default "";
 
   /**
    * Span description.